+++ R/misc.R
@@ -4668,7 +4668,21 @@ matrixShiftLongitude <- function(m, longitude) {
     if (max(longitude, na.rm = TRUE) > 180) {
         cut <- which.min(abs(longitude - 180))
         longitude <- c(longitude[seq.int(cut + 1L, n)] - 360, longitude[seq.int(1L, cut)])
-        m <- m[c(seq.int(cut + 1L, n), seq.int(1L, cut)), ]
+        #message("DEBUGGING matrixShiftLongitude()...")
+        #cat(vectorShow(length(dim(m))))
+        #cat(vectorShow(m))
+        #cat(vectorShow(n))
+        #cat(vectorShow(cut))
+        ndim <- length(dim(m))
+        if (identical(2L, ndim)) {
+            #cat("ndim is 2\n")
+            m <- m[c(seq.int(cut + 1L, n), seq.int(1L, cut)), ]
+        } else if (identical(3L, ndim)) {
+            #cat("ndim is 3\n")
+            m <- m[c(seq.int(cut + 1L, n), seq.int(1L, cut)), , ]
+        } else {
+            stop("can only handle 2D or 3D arrays, not a ", ndim, "D array")
+        }
     }
     list(m = m, longitude = longitude)
 }